+++ .gitignore
@@ -6,3 +6,4 @@ waveemapi.toml
 uploads
 data
 !tests/data
+lcov.info
